14/11 working on cifar

+++ models/import_models.sh
@@ -6,11 +6,36 @@ FCN32=./fcn-32s-pascal/deploy.prototxt
 #FCN8S_PROTO_URL=https://gist.githubusercontent.com/longjon/1bf3aa1e0b8e788d7e1d/raw/2711bb261ee4404faf2ddf5b9d0d2385ff3bcc3e/fcn-8s-pascal-deploy.prototxt
 FCN8S_PROTO_URL=https://github.com/HyeonwooNoh/DeconvNet/blob/master/model/FCN/fcn-8s-pascal-deploy.prototxt
 FCN8=./fcn-8s-pascal/deploy.prototxt
+GPU=./places_googlenet
 
 
 echo "downloading..."
+#for model in ${MODELS_file[@]}; do
+#    if [-f model];
+#    then
+#	echo "file $MODEL_FILE exist"
+#    else
+#	wget ${MODELS}
+if [ ! -f  "$FCN32" ]
+then 
+    wget $FCN32S_PROTO_URL -O $FCN32
+fi
 
-wget $FCN32S_PROTO_URL -O $FCN32
-wget $FCN8S_PROTO_URL -O $FCN8
+if [ ! -f "$FCN8" ]
+then
+    wget $FCN8S_PROTO_URL -O $FCN8
+fi
+
+if [ ! -d "$GPU" ]
+then 
+    echo "no $GPU"
+    mkdit $GPU
+    cd $GPU
+    wget http://vision.princeton.edu/pvt/GoogLeNet/Places/train_val_googlenet.prototxt -O deploy.prototxt
+    wget http://vision.princeton.edu/pvt/GoogLeNet/Places/places_googlenet.caffemodel 
+    cd ..
+fi
+
+#wget $FCN8S_PROTO_URL -O $FCN8
 
 echo "done"
